refactor: フォルダー名を適切な名称に修正
