+++ frontend/src/components/common/Chart.tsx
